move Exports method to export.go files
